removed all Grommet.io components and replaced with Styled component
