+++ src/components/Home.js
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import LoginBtn from './auth/LoginBtn'
 import LogoutBtn from './auth/LogoutBtn'
 import SignIn from './auth/SignIn'
-import SearchBox from './SearchBox'
+import SearchBox from './SearchBox/SearchBox'
 import Testimonials from './Testimonials'
 
 const ComingSoon = styled.h1`
